+++ lib/ts/recipe/accountlinking/types.ts
@@ -81,10 +81,7 @@ export type RecipeInterface = {
         users: User[];
         nextPaginationToken?: string;
     }>;
-    canCreatePrimaryUser: (input: {
-        recipeUserId: RecipeUserId;
-        userContext: UserContext;
-    }) => Promise<
+    canCreatePrimaryUser: (input: { recipeUserId: RecipeUserId; userContext: UserContext }) => Promise<
         | {
               status: "OK";
               wasAlreadyAPrimaryUser: boolean;
@@ -97,10 +94,7 @@ export type RecipeInterface = {
               description: string;
           }
     >;
-    createPrimaryUser: (input: {
-        recipeUserId: RecipeUserId;
-        userContext: UserContext;
-    }) => Promise<
+    createPrimaryUser: (input: { recipeUserId: RecipeUserId; userContext: UserContext }) => Promise<
         | {
               status: "OK";
               user: User;
@@ -139,11 +133,7 @@ export type RecipeInterface = {
               status: "INPUT_USER_IS_NOT_A_PRIMARY_USER";
           }
     >;
-    linkAccounts: (input: {
-        recipeUserId: RecipeUserId;
-        primaryUserId: string;
-        userContext: UserContext;
-    }) => Promise<
+    linkAccounts: (input: { recipeUserId: RecipeUserId; primaryUserId: string; userContext: UserContext }) => Promise<
         | {
               status: "OK";
               accountsAlreadyLinked: boolean;
@@ -163,10 +153,7 @@ export type RecipeInterface = {
               status: "INPUT_USER_IS_NOT_A_PRIMARY_USER";
           }
     >;
-    unlinkAccount: (input: {
-        recipeUserId: RecipeUserId;
-        userContext: UserContext;
-    }) => Promise<{
+    unlinkAccount: (input: { recipeUserId: RecipeUserId; userContext: UserContext }) => Promise<{
         status: "OK";
         wasRecipeUserDeleted: boolean;
         wasLinked: boolean;