+++ lib/ts/recipe/session/recipeImplementation.ts
@@ -13,7 +13,6 @@ import { buildFrontToken } from "./cookieAndHeaders";
 import { validateClaimsInPayload } from "./utils";
 import Session from "./sessionClass";
 import { Querier } from "../../querier";
-import NormalisedURLPath from "../../normalisedURLPath";
 import { JSONObject, NormalisedAppinfo, UserContext } from "../../types";
 import { logDebugMessage } from "../../logger";
 import { ParsedJWTInfo, parseJWTWithoutSignatureVerification } from "./jwt";
@@ -345,7 +344,7 @@ export default function getRecipeInterface(
                     : input.newAccessTokenPayload;
 
             let response = await querier.sendPostRequest(
-                new NormalisedURLPath("/recipe/session/regenerate"),
+                "/recipe/session/regenerate",
                 {
                     accessToken: input.accessToken,
                     userDataInJWT: newAccessTokenPayload,