+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## Unreleased
+
+-   Migrates backend-sdk-testing test to use a containerized core
+    -   Migrates from CircleCI to Github Actions
+-   Adds workflow to test supertokens-website
+    -   Updates `frontendIntegration` servers
+-   Fixes broken workflows (Example Tests, AWS Edge function compatibility)
+-   Sets up workflow to run auth-react tests
+    -   Updates test-servers to work with updated tests
+-   Fixes wrong types types from `string` to `number` on `createdAt`, `expiresAt` and `timeout` from the WebAuthn recipe
+-   Fixed `email` property by making it optional on `getGeneratedOptions` recipe implementation method
+
+## [23.0.0] - 2025-06-10
+
+-   Refactors internal logic of parsing cookies to check accessToken and optimizes it to avoid parsing unrelated cookies.
+-   The `getConsentRequest`, `acceptConsentRequest`, `rejectConsentRequest`, `acceptLoginRequest`, `rejectLoginRequest` and `introspectToken` can now possibly return an `ErrorOAuth2`.
+-   The `/oauth/introspect` can now possibly return an `ErrorAuth2`.
+-   The `User` class now has a `fromApi` function to normalize the user object returned from the API.
+-   Refactors querier to use dynamic request body and response body types inference.
+-   Refactor internal network calls made with querier to use the new dynamic types.
+-   Fixes an issue with fetch not supporting `cache: no-cache` in Cloudflare Workers.
+
 ## [22.1.0] - 2025-04-04
 
 -   Adds support for using `code_challenge_method` from OIDC provider response to determine whether to use PKCE or not.